requirements.txt
pandas==1.5.3
numpy==1.24.2
Pillow==9.4.0
jupyter==1.0.0
